Add lazy loading for COCO, fix some issues
