+++ dali/pipeline/operators/reader/coco_reader_op.h
@@ -26,45 +26,33 @@
 
 #include "dali/pipeline/operators/reader/reader_op.h"
 #include "dali/pipeline/operators/reader/loader/file_loader.h"
+#include "dali/pipeline/operators/reader/loader/coco_loader.h"
 #include "dali/pipeline/operators/reader/parser/coco_parser.h"
 
 namespace dali {
-
-void ParseAnnotationFilesHelper(std::vector<std::string> &annotations_filename,
-                                AnnotationMap &annotations_multimap,
-                                std::vector<std::pair<std::string, int>> &image_id_pairs,
-                                bool ltrb, bool ratio, float size_threshold, bool skip_empty);
 
 class COCOReader : public DataReader<CPUBackend, ImageLabelWrapper> {
  public:
   explicit COCOReader(const OpSpec& spec)
   : DataReader<CPUBackend, ImageLabelWrapper>(spec),
-    annotations_filename_(spec.GetRepeatedArgument<std::string>("annotations_file")),
-    ltrb_(spec.GetArgument<bool>("ltrb")),
-    ratio_(spec.GetArgument<bool>("ratio")),
-    size_threshold_(spec.GetArgument<float>("size_threshold")),
-    skip_empty_(spec.GetArgument<bool>("skip_empty")),
-    save_img_ids_(spec.GetArgument<bool>("save_img_ids")) {
-    ParseAnnotationFiles();
-
-    auto shuffle_after_epoch = spec.GetArgument<bool>("shuffle_after_epoch");
-    auto stick_to_shard = spec.GetArgument<bool>("stick_to_shard");
-
-    if (shuffle_after_epoch || stick_to_shard)
+    save_img_ids_(spec.GetArgument<bool>("save_img_ids")),
+    shuffle_after_epoch_(spec.GetArgument<bool>("shuffle_after_epoch")),
+    stick_to_shard_(spec.GetArgument<bool>("stick_to_shard")) {
+    if (shuffle_after_epoch_ || stick_to_shard_)
       DALI_ENFORCE(
-        !shuffle_after_epoch || !stick_to_shard,
+        !shuffle_after_epoch_ || !stick_to_shard_,
         "shuffle_after_epoch and stick_to_shard cannot be both true");
 
     if (spec.HasArgument("file_list"))
       loader_ = InitLoader<FileLoader>(
         spec,
         std::vector<std::pair<string, int>>(),
-        shuffle_after_epoch);
+        shuffle_after_epoch_);
     else
-      loader_ = InitLoader<FileLoader>(
+      loader_ = InitLoader<CocoLoader>(
         spec,
-        image_id_pairs_,
-        shuffle_after_epoch);
+        annotations_multimap_,
+        shuffle_after_epoch_);
     parser_.reset(new COCOParser(spec, annotations_multimap_, save_img_ids_));
   }
 
@@ -73,19 +61,11 @@ class COCOReader : public DataReader<CPUBackend, ImageLabelWrapper> {
   }
 
  protected:
-  void ParseAnnotationFiles() {
-    ParseAnnotationFilesHelper(annotations_filename_, annotations_multimap_,
-                                image_id_pairs_, ltrb_, ratio_, size_threshold_, skip_empty_);
-  }
-
-  std::vector<std::string> annotations_filename_;
   AnnotationMap annotations_multimap_;
-  std::vector<std::pair<std::string, int>> image_id_pairs_;
-  bool ltrb_;
-  bool ratio_;
-  float size_threshold_;
-  bool skip_empty_;
   bool save_img_ids_;
+  bool shuffle_after_epoch_;
+  bool stick_to_shard_;
+
   USE_READER_OPERATOR_MEMBERS(CPUBackend, ImageLabelWrapper);
 };
