+++ dali/pipeline/operators/reader/loader/indexed_file_loader.h
@@ -31,6 +31,8 @@ class IndexedFileLoader : public Loader<CPUBackend, Tensor<CPUBackend>> {
  public:
   explicit IndexedFileLoader(const OpSpec& options)
     : Loader(options),
+      uris_(options.GetRepeatedArgument<std::string>("path")),
+      index_uris_(options.GetRepeatedArgument<std::string>("index_path")),
       current_file_(nullptr) {
     }
 
@@ -109,11 +111,9 @@ class IndexedFileLoader : public Loader<CPUBackend, Tensor<CPUBackend>> {
   }
 
  protected:
-  void PrepareMetadata(const OpSpec& options) {
-    uris_ = options.GetRepeatedArgument<std::string>("path");
+  void PrepareMetadata() override {
     DALI_ENFORCE(!uris_.empty(), "No files specified.");
-    std::vector<std::string> index_uris = options.GetRepeatedArgument<std::string>("index_path");
-    ReadIndexFile(index_uris);
+    ReadIndexFile(index_uris_);
     DALI_ENFORCE(!indices_.empty(), "Content of index files should not be empty");
     current_file_index_ = INVALID_INDEX;
     Reset(true);
@@ -142,6 +142,7 @@ class IndexedFileLoader : public Loader<CPUBackend, Tensor<CPUBackend>> {
   }
 
   std::vector<std::string> uris_;
+  std::vector<std::string> index_uris_;
   std::vector<std::tuple<int64, int64, size_t>> indices_;
   size_t current_index_;
   size_t current_file_index_;
